Moved code for interacting with codetester api into separate package
