+++ src/main/kotlin/com/github/yniklas/intellijcodetesterupload/api/LoginException.kt
@@ -1,0 +1,3 @@
+package com.github.yniklas.intellijcodetesterupload.api
+
+class LoginException(message: String) : Exception(message)
