--- src/main/kotlin/com/github/yniklas/intellijcodetesterupload/services/ProjectService.kt
@@ -1,10 +1,0 @@
-package com.github.yniklas.intellijcodetesterupload.services
-
-import com.intellij.openapi.project.Project
-
-class ProjectService(project: Project) {
-
-    init {
-    }
-
-}
